ConstBitVec as (u128, usize) instead of (i128, i8)
Original-commit: https://github.com/mmcloughlin/veriisle-wasmtime/commit/26405115be2de5a1dab23bd5877d0835e42d3843

+++ cranelift/isle/isle/src/ast.rs
@@ -102,8 +102,8 @@ pub enum SpecExpr {
     },
     /// An operator that matches a constant bitvector value.
     ConstBitVec {
-        val: i128,
-        width: i8,
+        val: u128,
+        width: usize,
         pos: Pos,
     },
     /// An operator that matches a constant boolean value.
